service-kundali/requirements.txt
requests
numpy
absl-py
flask
pyswisseph==2.10.3.1